requirements.txt
tensorflow==2.15.0
python-telegram-bot==20.7
Pillow==10.1.0
numpy==1.24.3
matplotlib==3.8.2
scikit-learn==1.3.2
